Move table block "Show on map" toggle into the edit top bar. (#1903)
Keep the control always visible in edit mode so users can toggle map visibility before analysis results finish loading.
